ARROW-6212: [Java] Support vector rank operation
Given an unsorted vector, we want to get the index of the ith smallest element in the vector. This function is supported by the rank operation.

We provide an implementation that gets the index with the desired rank, without sorting the vector (the vector is left intact), and the implementation takes O(n) time, where n is the vector length.

Closes #5066 from liyafan82/fly_0812_rank and squashes the following commits:

623b08531 <liyafan82>  Support vector rank operation

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
Signed-off-by: Micah Kornfield <emkornfield@gmail.com>