ARROW-6216: [C++][Parquet] Expose codec compression level to user, add to Parquet writer properties
This patch allows the user to select a compression level to be used for when column data in Parquet is compressed. The user can select between better compression ratio or speed.

Reopening PR to supersede #5071 since the GitHub UI closed the PR when I attempted to push commits to the contributor's master branch

Closes #5192 from wesm/ARROW-6216-compression and squashes the following commits:

8f124619b <Wes McKinney> Remove GzipCodec ctor that's ambiguous in MSVC
beafdda49 <Wes McKinney> clang-format
1c905868c <Wes McKinney> Move default compression levels to headers, remove nullary ctors
06d8a14f3 <Wes McKinney> Use arrow::Compression in parquet:: namespace, address other stylistic nits
7f3bf2199 <Martin Radev> ARROW-6216:  Expose codec compression level to user

Lead-authored-by: Martin Radev <martin.b.radev@gmail.com>
Co-authored-by: Wes McKinney <wesm+git@apache.org>
Signed-off-by: Wes McKinney <wesm+git@apache.org>