ARROW-6556: [Python] Handle future removal of pandas SparseDataFrame
https://issues.apache.org/jira/browse/ARROW-6556

The plan in pandas is to remove SparseDataFrame/Series in pandas 1.0. By making sure our code works with that, we can ensure that a pandas release does not break the pyarrow release that is at that moment the latest stable release.
(and this also makes it easier for me to develop on master branches of both together)

This isn't merged yet in pandas, so we can also wait until that is done to merge this PR.
I was just trying out some things in pandas and saw that the pyarrow feather tests were failing when we remove those classes.

Closes #5377 from jorisvandenbossche/ARROW-6556-pandas-sparse and squashes the following commits:

705f0a725 <Joris Van den Bossche> ARROW-6556:  Handle future removal of pandas SparseDataFrame

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Wes McKinney <wesm+git@apache.org>