ARROW-6562: [GLib] Fix returning wrong sliced data of GArrowBuffer
We should not share the GBytes data of the original buffer with the
sliced buffer because the sliced buffer has different offset and size
with the original buffer.

The sliced buffer should have reference to the original buffer. If we
shouldn't have it, the original buffer may be freed when the sliced
buffer is used.

This change disables get_property of PROP_DATA. This is not related to
this fix. But I include this change. Sorry.

Users should use garrow_buffer_get_data() to get data instead of
get_property of PROP_DATA. So we should not provide get_property of
PROP_DATA.

Closes #5382 from kou/glib-buffer-fix-wrong-sliced-data and squashes the following commits:

7ebea8665 <Sutou Kouhei>  Fix returning wrong sliced data of GArrowBuffer

Authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Yosuke Shiro <yosuke.shiro615@gmail.com>