ARROW-6089: [Rust] [DataFusion]  Implement physical plan for "selection" operator
This PR implements the physical execution plan for the selection operator (the WHERE clause in a SQL query).

In order to have working tests, I also had to implement some subset of expressions (column reference, literal value, comparison expressions, and CAST). However, the goal of this PR is not to add complete support for all expressions but to implement the Selection operator. I will create separate JIRA/PRs for adding support for other expressions and data types in the physical query plan.

Closes #5320 from andygrove/ARROW-6089 and squashes the following commits:

6cad32726 <Andy Grove> Implement selection operator

Authored-by: Andy Grove <andygrove73@gmail.com>
Signed-off-by: Andy Grove <andygrove73@gmail.com>