ARROW-7092: [R] Add vignette for dplyr and datasets
The vignette could probably use some polishing, and it's not awesome that it's not a dynamic document (all of the code chunks and output are static, pregenerated) but I'm not sure of a better way since we can't include the example data in the package.

Feedback greatly appreciated. Given the push for the 0.16 release, I'll probably merge this as soon as CI is green because I have another issue (https://issues.apache.org/jira/browse/ARROW-7581) for documentation polishing for the release and I don't want this to block that. I can incorporate suggestions and edits folks have in that followup PR.

Closes #6222 from nealrichardson/dataset-vignette and squashes the following commits:

cca95f053 <Neal Richardson> Add notes on extension points for datasets
e3ba51daf <Neal Richardson> First pass at (static) dataset vignette; also add some print methods and various schema methods

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>