ARROW-7641: [R] Make dataset vignette have executable code:
This patch makes the dataset vignette executable, yet it also leaves in the expected output and conditionally shows it if the taxi data is not found locally. That way, the rendered vignette always looks useful.

In addition, the dataset print method is improved to report the number and format of the files it contains.

Closes #6247 from hadley/vignette-eval

Lead-authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Co-authored-by: Hadley Wickham <h.wickham@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>