ARROW-8167: [CI] Add support for skipping builds with skip pattern in pull request title
In certain cases, especially when someone adds a new build configuration it can be handy to disable all but that job.

Sadly the commit message is not available for the pull_request event, so we cannot have conditions on it, but we can use the pull request's title for this purpose.

Hopefully the pull request event payload will contain the head commit in the future.

Closes #6671 from kszucs/skip-ci

Authored-by: Krisztián Szűcs <szucs.krisztian@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>