ARROW-8644: [Python] Restore ParquetDataset behaviour to always include partition column for dask compatibility
Given that the original change (https://issues.apache.org/jira/browse/ARROW-3861 / https://github.com/apache/arrow/pull/7050) breaks dask's reading of partitioned datasets (it doesn't add the partition column to the list of columns to read, but expects it will still be read automatically), it doesn't seem worth it to me to fix this in the "old" ParquetDataset implementation.

But we can keep the "correct" behaviour in the Datasets API - based implementation going forward.

Closes #7096 from jorisvandenbossche/ARROW-8644-dask-partitioned

Authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: Wes McKinney <wesm+git@apache.org>