ARROW-8862: [C++] NumericBuilder should use MemoryPool passed to CTOR
`NumericBuilder` uses the `pool` (`MemoryPool*`) parameter to initialise the `ArrayBuilder` base class, but does not use it to initialise its own internal builder, `data_builder_` (`TypedBufferBuilder`).  For comparison `ArrayBuilder` uses the `pool` to initialise its own `null_bitmap_builder_` member (also a `TypedBufferBuilder`).

This effect was observed when trying to switch to a custom `MemoryPool` for performance reasons.  A hook was used to detect any use of the `MemoryPool` proved by `default_memory_pool()`.

Closes #7227 from sawatts/patch-1

Authored-by: Simon Watts <sawatts@users.noreply.github.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>