ARROW-8878: [R] try_download is confused when download.file.method isn't default
Originally reported #7058

`download.file()` can be called with different methods, mostly for legacy reasons according to the docs, but apparently some are on systems where it has been set to "wget" method. In classic R fashion, the behavior of the function changes significantly depending on how this global variable is set.

This patch should make the behavior more consistent regardless of the method used, so that the Linux installation won't proceed thinking a file was downloaded successfully when it in fact was not.

Closes #7304 from nealrichardson/download-file

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>